Organized the components
