+++ src/components/expenses/ExpenseItem.js
@@ -1,6 +1,6 @@
 import './ExpenseItem.css';
 import ExpenseDate from "./ExpenseDate";
-import Card from "./Card";
+import Card from "../ui/Card";
 
 export default function ExpenseItem({title, amount, date}) {
     return (
