+++ src/components/expenses/Expenses.js
@@ -1,5 +1,5 @@
 import ExpenseItem from "./ExpenseItem";
-import Card from "./Card";
+import Card from "../ui/Card";
 import './Expenses.css';
 
 export default function Expenses({items}){
